refactor(webapp): rename Meet Now modal files to camelCase
Rename the meetNowModal directory and its PascalCase file names to match
the agreed camelCase convention for paths introduced in this PR.
